Updated combined bank example using UUIDs.
